[GPU] Moved GPU remote headers to intel_gpu folder and namespace
