+++ src/inference/include/openvino/runtime/intel_gpu/ocl/va.hpp
@@ -7,7 +7,7 @@
  * shared Video Acceleration device contexts
  * and shared memory tensors which contain Video Acceleration surfaces
  *
- * @file openvino/runtime/gpu/ocl/va.hpp
+ * @file openvino/runtime/intel_gpu/ocl/va.hpp
  */
 #pragma once
 
@@ -18,7 +18,7 @@
 #include <memory>
 #include <string>
 
-#include "openvino/runtime/gpu/ocl/ocl.hpp"
+#include "openvino/runtime/intel_gpu/ocl/ocl.hpp"
 
 // clang-format off
 #include <va/va.h>
@@ -26,7 +26,7 @@
 
 namespace ov {
 namespace runtime {
-namespace gpu {
+namespace intel_gpu {
 namespace ocl {
 
 /**
@@ -147,6 +147,6 @@ public:
     }
 };
 }  // namespace ocl
-}  // namespace gpu
+}  // namespace intel_gpu
 }  // namespace runtime
 }  // namespace ov
